+++ src/utils/logger.py
@@ -3,7 +3,6 @@ Module to generate the logger.
 """
 
 import logging
-from constants import PROJECT_NAME
 
 
 def get_logger():
@@ -15,7 +14,8 @@ def get_logger():
     """
 
     # create logger
-    logger = logging.getLogger(PROJECT_NAME)
+    project_name = "kafgres"
+    logger = logging.getLogger(project_name)
     logger.setLevel(logging.DEBUG)
 
     # create console handler and set level to debug
